releases: set ref for github-workflow checkout steps (#26)
When using `actions/checkout` in a shared workflow call, the checkout
step will automatically use the ref for the caller repo (ie v0.3.2 for
baton-xyz), but not for this workflow repo - instead defaulting to
`main`.

example:
- in `baton-xyz`, cutting a new release for `v0.3.2` calls the shared
workflow: `uses:
ConductorOne/github-workflows/.github/workflows/release.yaml@v2`
- it uses the `release.yaml` file in `ConductorOne/github-workflows`
based on the `v2` tag reference
- but the `releases.yaml` then calls to checkout other files in it's own
repo `github-workflows`, which default to `main`

As a result, for all past versions (v1, v2, and the current v3) of the
shared release workflow, the workflow is using the goreleaser and gon
files from `main`.

This PR breaks that pattern so that once it's merged and tagged for the
floating `v3`, all future versions will work as expected. It also means
the following files cannot be changed unless they are backwards
compatible with prior `release.yaml` versions. Breaking changes will
require new files added.
- `.gon-amd64-template.json`
- `.gon-arm64-template.json`
- `.goreleaser-docker-template-no-lambda.yaml`
- `.goreleaser-docker-template.yaml`
- `.goreleaser-template.yaml`