separate docker OCI and docker lambda (#36)
separates out the steps and config files for docker OCI and docker
Lambda.

in a followup PR we'll be pushing OCI docker images to a public ECR so
we can't use multiple aws roles in a single config.

also moves some templates to a `/templates` folder. in the future we'll
depreciate/remove releases for v1-v2 so that we can clean up existing
the top level template files.